Fix crash when using BottomNavigationView
BottomNavigationView is contained in a support library using which
requires an AppCompat theme, but the project is not based on that
theme. Therefore, trying to use the component was causing the app
to crash.

To fix it, another theme based on AppCompat has been added and the
particular activity containing the BottomNavigationView has been
specified to use that theme (in the manifest file). The rest of the
activities will continue using the original theme.