Fix crash on tapping Profile tab
App was crashing with the error message "lateinit property homePage
has not been initialized".

That's because the bindWidgets() method was never called upon
activity instantiation. Then in mOnNavigationItemSelectedListener(),
when R.id.navigation_profile was selected, homePage.visibility couldn't
be accessed, because bindWidgets() hadn't been called before.

Rather than calling bindWidget() everytime a tab is selected, we just
call it once in onCreate() and fix the issue.